fix(m2): greenfield EF persist and clear Sonar new violations
Nothing is deployed, so drop the ADO.NET compatibility path and initialize from the EF model. Ban raw SQL in Core and name the sync concurrency cap so the PR gate is not blocked on S1067, S1541, or S109.

+++ CipherBank-app.Core/Persist/Entities/RateSnapshotEntity.cs
@@ -8,9 +8,9 @@ public sealed record RateSnapshotEntity
 {
     public string Symbol { get; set; } = string.Empty;
 
-    public double Usd { get; set; }
+    public decimal Usd { get; set; }
 
-    public double Change24H { get; set; }
+    public decimal Change24H { get; set; }
 
     public long UpdatedAtMs { get; set; }
 }
